+++ packages/video_player_videohole/tizen/src/video_player.h
@@ -18,7 +18,6 @@
 #include <string>
 #include <utility>
 
-#include "ecore_wl2_window_proxy.h"
 #include "ffi_messages.h"
 
 namespace video_player_videohole_tizen {
@@ -85,7 +84,6 @@ class VideoPlayer {
 
   int64_t player_id_;
   std::mutex queue_mutex_;
-  std::unique_ptr<EcoreWl2WindowProxy> ecore_wl2_window_proxy_ = nullptr;
   flutter::BinaryMessenger *binary_messenger_;
   bool is_initialized_ = false;
   FlutterDesktopViewRef flutter_view_;