+++ packages/video_player_videohole/CHANGELOG.md
@@ -1,10 +1,14 @@
 ## 0.6.0
 
-* Resolve `strict_top_level_inference` lint.
 * Migrate from Platform Channels to Dart FFI.
 * Replace EventChannel with FFI port for event callbacks.
 * Add JSON serialization for complex parameters.
 * Add pending seekTo handling.
+
+## 0.5.11
+
+* Replace ecore-wl2 code with tizen window manager plugin.
+* Resolve `strict_top_level_inference` lint.
 
 ## 0.5.10